+++ tests/imephu/salt/annotation/test_telescope/test_salticam_slot_annotation_position_angle0_.png
